enable reconstructing opendp polard pipaline in client for archives
