+++ client/lomas_client/client.py
@@ -6,7 +6,7 @@ import pandas as pd
 import polars as pl
 from fastapi import status
 from opendp.mod import enable_features
-from opendp_logger import enable_logging, make_load_json
+from opendp_logger import enable_logging
 from opentelemetry.instrumentation.logging import LoggingInstrumentor
 from pydantic import ValidationError
 
@@ -23,16 +23,14 @@ from lomas_client.models.config import ClientConfig
 from lomas_client.utils import raise_error, validate_model_response_direct
 from lomas_core.constants import DPLibraries
 from lomas_core.instrumentation import init_telemetry
-from lomas_core.models.requests import (
-    GetDummyDataset,
-    LomasRequestModel,
-)
+from lomas_core.models.requests import GetDummyDataset, LomasRequestModel, OpenDPQueryModel
 from lomas_core.models.responses import (
     DummyDsResponse,
     InitialBudgetResponse,
     RemainingBudgetResponse,
     SpentBudgetResponse,
 )
+from lomas_core.opendp_utils import reconstruct_measurement_pipeline
 
 # Opendp_logger
 enable_logging()
@@ -234,8 +232,10 @@ class Client:
                         else:
                             query["response"]["result"] = pd.DataFrame(res)
                     case DPLibraries.OPENDP:
-                        opdp_query = make_load_json(query["client_input"]["opendp_json"])
-                        query["client_input"]["opendp_json"] = opdp_query
+                        query_json = OpenDPQueryModel.model_validate(query["client_input"])
+                        query["client_input"]["opendp_json"] = reconstruct_measurement_pipeline(
+                            query_json, self.get_dataset_metadata()
+                        )
                     case DPLibraries.DIFFPRIVLIB:
                         model = base64.b64decode(query["response"]["result"]["model"])
                         query["response"]["result"]["model"] = pickle.loads(model)
