+++ server/lomas_server/constants.py
@@ -1,10 +1,6 @@
 import string
 from enum import StrEnum
 
-from opendp import measures as ms, typing as tp
-
-from lomas_core.constants import OpenDpMechanism
-
 # Misc
 # -----------------------------------------------------------------------------
 
@@ -83,18 +79,3 @@ class OpenDPDatasetInputMetric(StrEnum):
     HAMMING_DISTANCE = "HammingDistance"
 
     INT_DISTANCE = "u32"  # opendp type for distance between datasets
-
-
-OPENDP_TYPE_MAPPING = {
-    "int32": tp.i32,
-    "float32": tp.f32,
-    "int64": tp.i64,
-    "float64": tp.f64,
-    "string": tp.String,
-    "boolean": bool,
-}
-
-OPENDP_OUTPUT_MEASURE: dict[OpenDpMechanism, tp.Measure] = {
-    OpenDpMechanism.LAPLACE: ms.max_divergence(),
-    OpenDpMechanism.GAUSSIAN: ms.zero_concentrated_divergence(),
-}
